src/main/java/kmichalski/scoreboard/service/BoardService.java
package kmichalski.scoreboard.service;

import kmichalski.scoreboard.model.Game;
import kmichalski.scoreboard.model.GameStatus;
import kmichalski.scoreboard.model.Team;
import kmichalski.scoreboard.repostiory.GameRepository;
import kmichalski.scoreboard.repostiory.TeamRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class BoardService {
    private final GameRepository gameRepository;
    private final TeamRepository teamRepository;

    public Game createNewGame(Team homeTeam, Team awayTeam) {
        Team savedHomeTeam = teamRepository.save(homeTeam);
        Team savedAwayTeam = teamRepository.save(awayTeam);
        Game newGame = Game.builder()
                .gameStatus(GameStatus.NEW)
                .homeTeam(savedHomeTeam)
                .awayTeam(savedAwayTeam)
                .awayTeamScore(null)
                .homeTeamScore(null)
                .build();
        return gameRepository.save(newGame);
    }
}
